Exception handling - delete

+++ src/main/java/com/project/study/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.project.study.entities.User;
 import com.project.study.repositories.UserRepository;
+import com.project.study.services.exceptions.DatabaseException;
 import com.project.study.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -24,10 +27,19 @@ public class UserService {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
+	
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
+	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException  e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 }
